Add files via upload
A simple demonstration of calculating and displaying the result with a couple of integer inputs from the user. No attempt has been made to validate the inputs, however, you may learn how to calculate power of a number here. 

The program would give the following errors:
1. ValueError - if enter is hit without providing a number as input
2. ValueError - if float numbers are provided as inputs
3. ValueError - if character values are entered

We will learn how to deal with this in forthcoming examples. For now you may work with non-float integer values to run this.