Add files via upload
User is asked to input numbers. The entries are not validated, however, if non-float numbers are provided, the program calculates the sum of the current and previous entries and displays the total on each entry. Input is not validated, however, you are introduced to the concepts  of using:
1. user input capture
2. while loop
2. type casting
3. increment of numbers by 1 and other values 
4. print results on the shell console

A sample of valid input and the output on IDLE shell is given below:
Input num(1): 2
Sum of current and previous input(s) is:  2
Input num(2): 5
Sum of current and previous input(s) is:  7
Input num(3): 7
Sum of current and previous input(s) is:  14
Input num(4): 4
Sum of current and previous input(s) is:  18
Input num(5): 8
Sum of current and previous input(s) is:  26