Add files via upload
This program is not based on any real world scenario. Here we focus on editing of the program interface as may be viewed on IDLE shell - no attempt has been made to validate the inputs. This is more like the editing done by developers working on legacy system. You may be interested to check out the 'nitty-gritties' involved. The testing has been done with least possible ideal values of '1' for all entries and '9999999' as a max for all entries to adjust the editing attempted. 

A sample of the output as on IDLE shell is as follows (the actual output is tidy when viewed on the shell):
---------------------------------------------------------------------------
#                      Calculation of Monthly Profit                      #
---------------------------------------------------------------------------
Sold Quantity.........: 1
Raw Material Cost.....: 1
Manufacturing Cost....: 1
Profit Margin (in%)...: 1
Tax Deduction (in%)...: 1
---------------------------------------------------------------------------
Cost of product (=raw material + mfg. cost) is...:  2
Selling price (=CoP+Prft_mrgn/100 * CoP) is......:  2.02
Net sales (=Sold_qty * Selling_Price) is.........:  2.02
Net CP (=Sold_qty * CoP) is......................:  2
Gross Profit (=net_sales-net_cost_price) is......:  0.020000000000000018
---------------------------------------------------------------------------
Monthly Profit of Organization (=gp-td/100*gp)...:  0.01980000000000002
---------------------------------------------------------------------------