Add files via upload
This program prints even characters in an input string. It also demonstrates:
1. Defining a function and using the same
2. Conversion of input string to a list
3. Use of \n when printing
4. Using for loop on a list

A sample of execution of the code on Python IDLE is as follows:
Enter a string: Hi! I am Thomas but not Thomas Edison

Output of function spliting intput string to a list... ['H', 'i', '!', ' ', 'I', ' ', 'a', 'm', ' ', 'T', 'h', 'o', 'm', 'a', 's', ' ', 'b', 'u', 't', ' ', 'n', 'o', 't', ' ', 'T', 'h', 'o', 'm', 'a', 's', ' ', 'E', 'd', 'i', 's', 'o', 'n']


The characters in the even spaces are:  i
The characters in the even spaces are:   
The characters in the even spaces are:   
The characters in the even spaces are:  m
The characters in the even spaces are:  T
The characters in the even spaces are:  o
The characters in the even spaces are:  a
The characters in the even spaces are:   
The characters in the even spaces are:  u
The characters in the even spaces are:   
The characters in the even spaces are:  o
The characters in the even spaces are:   
The characters in the even spaces are:  h
The characters in the even spaces are:  m
The characters in the even spaces are:  s
The characters in the even spaces are:  E
The characters in the even spaces are:  i
The characters in the even spaces are:  o