Add files via upload
This program prints the count of a chosen character in an input string. It also demonstrates the use of count() function on a string. 

Sample output of the code execution on Python Shell:
----------------------------------------------------------------------
Enter a string: I am not an Australian
Enter a char to count: e
Count of e in the input string is:  0
----------------------------------------------------------------------
Enter a string: Two and two makes four
Enter a char to count: O
Count of O in the input string is:  0
----------------------------------------------------------------------
Enter a string: Let's get some pasta for dinner
Enter a char to count: e
Count of e in the input string is:  4
----------------------------------------------------------------------
Enter a string: 
Enter a char to count: 
Count of  in the input string is:  1
----------------------------------------------------------------------