Merging Dictionaries
This program demonstrates two methods to merge two or more dictionaries. One uses the update() function while the other uses the extend() function. You may want to make note of how the for-loop is used to merge with extend() function.  

Here is the output on Python Shell when the program is executed:
---------------------------------------------------------------------------------------------------------
Method1: merge dictionaries with update()
dic1 is:  {1: 20, 2: 20}
dic2 is:  {3: 30, 4: 40}
dic3 is:  {5: 50, 6: 60}
dic1.update(dic2) is dic1 merged with dic2 is:  {1: 20, 2: 20, 3: 30, 4: 40}
dic1.update(dic3) is dic3 merged with modified dic1 is:  {1: 20, 2: 20, 3: 30, 4: 40, 5: 50, 6: 60}
dic1 is now:  {1: 20, 2: 20, 3: 30, 4: 40, 5: 50, 6: 60}


Method2: merge by making a list of dictionaries
dic4 is:  {'a': 70, 'b': 20}
dic5 is:  {'c': 30, 'd': 40}
dic6 is:  {'e': 50, 'f': 60}
dic7 is:  {'g': 20, 'h': 20}
dic8 is:  {'i': 30, 'j': 40}
dic9 is:  {'k': 50, 'l': 60}
{'a': 70, 'b': 20, 'c': 30, 'd': 40, 'e': 50, 'f': 60, 'g': 20, 'h': 20, 'i': 30, 'j': 40, 'k': 50, 'l': 60}
---------------------------------------------------------------------------------------------------------