Demo of Try-Except-Else_Finally
This program does calculation on a range starting from 1 and the upper limit as provided by the user. Stores the calculation in a dictionary and displays the same with the help of a Try-Except-Else_Finally exception handler.

A sample of the trial execution of the code on Python IDLE Shell is given below:
-------------------------------------------------------------------------------------------------------------------------------
Input a number: 6.8
6.8
{1.0: 1.0, 1.5: 2.25, 2.0: 4.0, 2.5: 6.25, 3.0: 9.0, 3.5: 12.25, 4.0: 16.0, 4.5: 20.25, 5.0: 25.0, 5.5: 30.25, 6.0: 36.0, 6.5: 42.25}
Change program logic only if you expected a different output.
Do you wnat to re-run the program (y/Y)?: y
y
Input a number: 3.89
3.89
{1.0: 1.0, 1.5: 2.25, 2.0: 4.0, 2.5: 6.25, 3.0: 9.0, 3.5: 12.25, 4.0: 16.0, 4.5: 20.25, 5.0: 25.0, 5.5: 30.25, 6.0: 36.0, 6.5: 42.25}
Change program logic only if you expected a different output.
Do you wnat to re-run the program (y/Y)?: n
n
Hope you liked the demo! Do come back for more.
-------------------------------------------------------------------------------------------------------------------------------