Update Py_IdentifyIfTwoInputStringsAreIdentical.py
Below is the output of two sample execution of the code:
--------------------------------------------------------------------------------------------------------------------------------------
Enter a string: "Be strong and courageous. Do not be afraid or terrified because of them, for the LORD your God goes with you; he will never leave you nor forsake you."
Enter another string: Deuteronomy 31:6 "Be strong and courageous. Do not be afraid or terrified because of them, for the LORD your God goes with you; he will never leave you nor forsake you."
Strings are not identical
--------------------------------------------------------------------------------------------------------------------------------------
Enter a string: Isaiah 40:31 "But they who wait for the Lord shall renew their strength; they shall mount up with wings like eagles; they shall run and not be weary; they shall walk and not faint."       
Enter another string: Isaiah 40:31 "But they who wait for the Lord shall renew their strength; they shall mount up with wings like eagles; they shall run and not be weary; they shall walk and not faint." 
Identical strings entered!
--------------------------------------------------------------------------------------------------------------------------------------